+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Current
 
 - A proper git tag name is shown as the version number in footer (if available, otherwise the commit hash is used as it was before)
+- Better response if a user tries to see the details of someone else's alert (https://github.com/riparias/pterois/issues/223)
 
 # v1.0.0  (2023-07-12)